Remove remote spaces
This patch completely removes the concept of remote spaces in Juju. We
only had this for binding offers on different models, but it is not
useful since none of the providers support `ProviderSpaceInfo` and we
have removed the `AreSpacesRoutable` on the providers.
The cleanup includes the checks on the application offers API.