Update controlsocket to user userservice
The controlsocket used the legacy state package to add and remove
users. This swaps over the origin of the user service from mongo
to dqlite. Permissions aren't migrated, as that still requires
the mongo database. Once we transitioned away from legacy state
for permissions this should be updated.

Also I swapped out the fake/stub mocks in the test for real mocks.