Remove controller conifg from state
Now that we've removed the dependency of the dbaccessor on the
peergrouper we can now finish off the controller config. By removing
the state methods we can now see the full extent to which the implementations
need to be updated.

This will be a rather large PR as we just need to see it gone.