Fix a regression in charm deploy
A fix was previously made here https://github.com/juju/juju/pull/16692
to increase compatibility with 2.9 controllers and 3.x clients by adding
a compatibility layer

However, there is an edge case with charm url which we see when
deploying some local charms with pylibjuju, causing deploys to be
impossible

Sometimes pylibjuju uploads local charms in such a way that the
resulting charm url in Mongo does not include a series. Since the
compatibility layer ensures a series is present in curls when we deploy,
this charm becomes impossible to deploy

We considered a few approaches to this. The best is to drop this
mutation to the charm url, restoring the behavior partially to 2.9.46
Series in mongo charm urls, even in 2.9 is a very legacy feature which
we do not think is actually required anywhere.

However, the consequences of this need to be thoroughly investigated in QA
steps

Resolves: https://bugs.launchpad.net/juju/+bug/2058311