Add new facades version for remote space removal
This patch creates new facades version for the aplication,
aplicationoffers and crossmodel facades which make use of the RPC
structs that get the remote spaces (and bindings) removed.

This is an intermediate step before the complete removal in 4.0 and no
longer supporting remote spaces at all.

See https://github.com/juju/juju/pull/17056