Add permission access to add user args
Adding permissions to users when adding a user allows us to setup
permissions easily and conveniently. The problem is that permissions
aren't modelled in a nice way. Consider;

 - Read, Write, Consume and Admin are model access
 - Login, AddModel and Superuser are controller access

We should prevent Access from being a blanket string, instead it
should be a tree of types.

    type Access string
    type ModelAccess Access
    type ControllerAccess Access

Then we can add:

    func (a ModelAccess) Access() Accesss

This shouldn't be the final form, but it's a better representation
of the type access.