Pass controller and model txn runners in bootstrap
During bootstrap we need to access both the controller (global db)
and the model (controller model db). To prevent the need for hacks
to get that information, this revisits the bootstrap concerns and
moves back to bootstrap options. The options take the controller and
the model. Thus you can access any part of each model during bootstrap.