Refactors credential ID to Key.
We had previously introduced a type for the credential natural key of
cloud, name and owner and named this id. However this is more of a key
in our system being a natural way for users to refer to credentials.

ID should be reserved for the unique id we assign the credential in the
controller. This PR makes way for the credential ID to become it's own
type.