Wire up create model info into the model manager
In order to have the correct model info in the model database we
need to get the service factory for the correct model and then
insert the model info.

The read-only data can correctly live in the model database for
easy access. As this data is immutable we can get a lot of value
from this for things like the provider/broker workers.