Fix a regression when deploying with certain clients
We found that certain old clients (juju 2.8, pylibjuju 3.0) would deploy
an application without a base or series in the charm origin. Instead,
the series would only be provided with the 'series' param (and the
charm url)

In this case, we would attempt to parse the base in the origin,
expecting it to be there, and fail.

We resolve this by taking another approach. We assume nothing about the
arg provided, look for all the series provided, and return them if
they're equal

This ensure that no matter what client we use, so long as it sends an OS
(if it doesn't there's nothing we can do, this is a legitmate failure)
and if multiple series/bases are present, so long as they match
(non-matching series would also be a legitmate failure)