Remove check for validating controller config spaces
When passing a controller config, if the space is provided, we verify
that the controller machines do actually have at least one address in
the provided space.

We remove this check, which makes use of the legacy spaces domain, and
this should be reimplemented when we finish the migration of controller
config to dqlite.