DDL changes for Juju controller models.
This change updates the Juju DQlite DDL so that models are now in their
own table. We have done this as it does not make a great deal of sense
to have them split over 2 tables as per the original design.

This change also adds a model finalise flag to the table allowing models
to be created without the need for a transaction to do the whole
operation.

This commit also brings in changes around how we track DQlite
namespaces. Previously we use a model list table to track the dqlite
namespaces a controller managers. Now we use a dedicated namespace table
to track these namespaces.

The advantages of performing namespace tracking this way is that we are
able to seperate the lifecycle of a namespace and a model and perform
cleanup of the namespaces after the fact.