DDL changes for Juju controller models.
This change updates the Juju DQlite DDL so that models are now in their
own table. We have done this as it does not make a great deal of sense
to have them split over 2 tables as per the original design.

This change also adds a model finalise flag to the table allowing models
to be created without the need for a transaction to do the whole
operation.