Rename user to access
Domains are not entities. We should stop drawing lines between
services because they act and feel like entities. We're leaving that
old concept behind. Users and permissions are access controls over
a given aspect of juju. If we feel like permissions are a different
domain in the future, we should split them up, but right now they
are one and the same domain. At that point we will have a better
understanding of domains, permissions and users. I suspect though
that authn and authz are unfortunately conflicted. We should do
our best that the DB is not. At the moment the two concerns are
separate in the database and that's good enough (for now). If we
want to implement a proper authz and authn control at the API
server level, I feel that the access package should still suffice.

This removes the need to share state between users and permissions.
Erradicating the problem of how we share state between domains.