Use c.Check in application unit tests
Instead of c.Assert where failing the check will not impact code further
on. We built up a bad habit of using c.Assert un-necessarily. This stops
test execution when it fails, meaning we do not see future failures.

Assert an error is nil before Check on it's value. In a table
driven test, Check the error inside of an if statement containing a
Check on the value. Therefore the test will continue, but not panic on
check a value which does not exist, failing at the end.