Add support for Pebble change-update notice
This is basically just adding the notice type in the right places,
with the effect that we'll now respond to Pebble change-update notices
by sending a pebble-change-updated event to the charm.

This PR also updates "sidecar" integration tests and the pebble-notices
test charm with a change-update notice test.

For reference, Pebble Notices support (for the "custom" notice type)
was originally introduced in PR #16428.