Remove checks on spaces bound to endpoints when deleting spaces
On the environs, when we reload spaces at bootstrap time, we remove the spaces that are
not present in the substrate. Before doing this we check if the spaces
are bound to an endpoint. This check is useless because it's only used
at bootstrap time.

Nonetheless, if we were to check this, we must do it using the
relational integrity that the new dqlite schema provides, so we should
have a foreign key violation when trying to delete a space used in
endpoint bindings.