Eventsource multiwatcher
The multiwatcher was originally a notifywatcher, so it emitted changes
for type struct{}. This isn't actually required, we could just
generalize it for type T. This moves the multi watcher from
core watcher to eventsource. Eventually everything in core/watcher
will be replaced with core/watcher/eventsource (just not yet).