Wire spaces and subnets client facades
This patch wires the spaces and subnets client facades with the new
dqlite network domain, getting rid of the usages of the legacy mongodb
state regarding spaces and subnets.

Also, two methods (`UpdateSpace` and `GetAllSubnets`) where added to the
spaces and subnets services since they are needed in the facades.