Include sql in verbose mode for tests
Whilst debugging a test failure, including the SQL being executed
helped diagnose the problem. Ideally, we would add a new flag for
everyone to check against, but that requires a lot of reworking of
the tests. Instead, for now we can just set an envvar whilst running
the tests and it spew out what it ran and how it got stuck.