Local DB with schema
When debugging or wanting to play with the db schema, it's handy to
be able to apply the db to a given dqlite instance. The following
does this for controllers and models.

For experimenting with the domain model this is really useful.

I've removed the infer schema, as I don't believe we need this anymore.
It's not maintained and it served it purpose. If we do want it, we can
use the older branches to get it.