Fix subnet retrieval on space when empty subnets
When a space has no subnets, then the subnets field of the space struct
is filled with one empty subnet (every field included the ID is empty,
only the spaceName is set to the correct space). This is wrong, and it
happens because of how we construct the space struct when we fetch no
subnets and we serialize it onto the space.

This patch fixes this and adds one check to the tests.