cancelable API calls
If a watcher is incorrectly coded and refuses to stop _after_ the
initial changes are sent (strings and notify watchers), and the
Next call doesn't return any information, then the watcher will never
stop. This manifested itself during a migration when the secrets
pruner refused to stop because it used a testing watcher in production
code and didn't do the right thing.

The following ensure that we can give up and the client won't be
broken for ever.