Unblock model migrations
During model migrations to juju 4.0, CreateApplication will sometimes be
called on the service layer.

This domain is still heavily WIP. Calls to CreateApplication will not
include a specific charm in their params. At the moment, this charm is
only used to create storage.

Implement a temporary protective solution to unblock migrations so work
elsewhere can continue.