Generate triggers
Triggers were written whereby any update to the database would
cause a new insert into the change log, even if the value was
identical. This causes a lot of churn in the change log.
Instead we can generate the triggers using the columns from the
actual database schema. The triggers will always be correct and
always up to date.