Update SQL queries.
With finding the uuid of the user's creator, as it cannot be NULL, a
simpler JOIN can be used.

Updated the delete query to be cogent - as the permission is not needed
to find the uuid of the user from the name.

Made similar updates to the update permission query as well. Though to
determine disabled, the v_user_auth is needed.