Dqlite repl dump to sqlite3
A lot of sqlite cli commands[1] are not available in the dqlite cli.
Adding the ability to .dump to sqlite3 files then allows us to use
sqlite3 cli directly on the files.

The main change piggy backs on to existing infrastructure that's
available. The only problem is that we do need a real dqlite
lib to be installed. If that's not available we can piggy back on
to the static lib that's available.

 1. https://sqlite.org/cli.html