String-hack charm url in legacy handler
This is required since old controllers do not allow charm urls to change
during migration, and upload charms with a broken up charm url including
a series

As such, this handler needs to be able to reason with charm urls with
series. This is the only place in Juju this is required, so instead of
continuing to support series, string-hack the charm url to include a
series for compatibility.

This can be dropped when we no longer support migration from juju 3.3