Remove dump options and use unix socket
We really, really do not want people to use this dqlite-repl on live
production databases. The sole purpose for this command is for
interactive development. As dqlite can not provide some of the facilities
that sqlite can, we sometimes need to drop into sqlite proper.

Using a unix socket prevents anyone from accidently using the real
dqlite server without modifications. All the files for the dump will
use the db-path that's passed in. If nothing is suggested then a
temp directoy will be created.