Add permission sql to cloud domain.
When creating a cloud, include the user name to also add a permission
row for the user to have Admin permissions on their new cloud. Impacts
bootstrap and create cloud. There is overlap with UpdateCloud as well,
however unless a cloud is being created no user name is necessary. Cloud
permissions are not changed on update cloud.