Restore runcmd shell flag after setting juju's runcmd
Juju uses `set -xe` to force its runcmds to cause the script to fail
when an individual command fails. This shell option is not unset prior
to postruncmd commands, which leaks a side-effect to users. This causes
postruncmd scripts to behave differently than runcmd does and is not
documented.

Fix it.

LP: 1978454