Add spaces reload functionality on network domain
This patch adds the reload spaces functionality (which still exists on
the environs) to the new dqlite network domain. This was only possible
thanks to the work introduced by the provider spaces in https://github.com/juju/juju/pull/17165.

The only changes outside the scope of the network domain, are the logger
interfaces on the service factory, which has been updated with the Info
and Trace levels.