Use uuid v7 when adding a space, change method signature
This patch updates the google/uuid library so we can make use of the new
UUID V7 support. More specificaly, when adding a space we check if the
provided space contains an ID, if it doesn't we generate a new V7 UUID
for it.

We also changed the signature on the AddSpace method of the service
layer to do this.