Move caas provider package
The caas package should be located in the provider package. These
providers can all live together. This will prevent the special
case of the caas provider and make it feel like a real provider.

This is an ongoing effort to clean up the root of the juju repo.
As nobody should be depending on the caas provider directly, so we
should be free to move this.