Refactor UpsertCloud into CreateCloud and UpdateCloud.
The term upsert is a db concept and should not leak from the domain
state layers.

For clouds specifically, create and update are different as only create
will handle the permission of the owner for the cloud. It's easier to be
specific about which is happening rather than having to guess and maybe
provide the owner's name.