Delete DB contexts
As we can't drop a database, when a migration fails, we have to
delete the contents of the database. This includes all the schema
that was added to the database. The rationale for deleting the schema
is that, if you migrate src:A to dst:A, but that fails and then you
perform a upgrade to dst via a patch release, there isn't a guarantee
that dst:A schema is completely updated. So removing everything leaves
us with a blank state.

There have been thoughts around aliasing tables instead of a clean
slate, but this is fraught with danger if you accidently write to
the wrong DB (i.e. the alias one).

In addition to this, we also can't clear the DB in one query, as we
get a segfault. The code batches up the statements, so we're not running
a query for every statement, but should be enough to prevent a
segfault.

The way that we drop everything from the db needs to happen in this
order to prevent constraint violations. In addition we ensure that
the foreign key param is set to _false_ to also prevent any issues
in ordering of deletions.