Remove check on constraints before deleting a space.
On the reload space functionality, the check to make sure a space isn't
used in a constraint has been removed until we migrate constraints to
dqlite.
This goes in the same direction as what has been done for endpoint
bindings, and a TODO has been added for both checks. These will be added
again once we migrate constraints/bindings to dqlite.