Ensure that we stop the worker upon deletion
The worker needs to be stopped upon deletion. We want to ensure it's
dead. We don't want to allow the worker to spring back up. So having
a sentinel error for when we want to delete a model propergate through
the stack allows others to know when to give up.