Introduces cloud and credential loading on id.
We have logic outside of the respective cloud and credential state
domains that require loading cloud and credentials for a model. This
logic is complicated and error prone.

To deal with this this commit introduces new views for cloud and
credential that makes the sql for loading respective values easier. It
also provides state methods that can be resused from other states for
loading by id.

Secret backend can now get the id values for a model and call the
associated load values without any bugs and correct behaviour.