Make model deletion optional
Until we fully implement tearing down of a model, we can't currently
remove a model until the very last moment. This code just makes it
optional, so that we can state that during a model migration we do
want to delete the model so we can retry.

In addition, I've also fixed the dbaccessor to correctly terminate
transactions based on if the tomb is dying. See context sourceable.