Update subnet ddl to remove constraint on provider network id
This commit removes the unique constraint on the subnet's provider network id.
This error was shown when deploying on dual-stack lxd, where we have two
subnets for the same provider network id.
Also, the Network service will take as inputs both a provider and a
broker and decide which one to use based on a query to the model table,
where we retrieve the cloud type.