Add model last login table and functions
Add a new table to the controller schema to keep track of the last user
to log in to each model.

Update the UpdateLastLogin function to record this and take the model
uuid as an arguement.

Add a function LastModelLogin to get the last user to log in to a model.