Write read-only model after finalise
Until you finalise the model, the model type can not be read. This
is what we want for most if not all circumstances. Yet we're in a
situation that we haven't fully migrated the read-only model to where
we would like. Until that time, write the read-only model after the
finaliser has been called.

This will have to be reverted at a later date, but will prevent
Juju from being broken.