Add watcher for subnets
This patch adds a new subnets watcher on the network domain. This
watcher is formed by composing one watcher out of two: one for the
subnet table and the second one for the subnet_association table. The
second one is needed to watch underlay fan subnets.

This watcher is wired on the apiserver for its usage on the firewaller
worker.