Custom initial namespace query
The initial namespace query was originally a string, but that
prevents creating custom sql that requires things like where et al.

To provide a better way to do that, we create a HOF (high order
function) to pass in a default Query, with the ability to supply your
own.

The concept is really simple.