Drop GetSeriesFromBase call from equinix provider
It looks like this call was added as a mistake. Examining the history of
these lines, before we called GetSeriesFromBase, we called
UbuntuSeriesVersion with the series param on InstanceConfig

Reading the error from this function, as we used to, was a way determine
if a series was an ubuntu series.

Now that we use bases, we should instead just check the OS attribute of
the base