Remove LatestLTS & LatestLTSBase
These function both rely on distro-info sniffing which is to be removed

They were both used oth these functions are used only in tests with one
exception.

LatestLTSBase was used in the BaseSelector as a default base to deploy
charms to, if none are provded via flag or default-base model-config

We wish to remove this behaviour because it can lead to unexpected
changes in behaviour when deploying a charm within the same version of
Juju.

The direction we are moving in with workloads is to hard code a list of
supported bases inside Juju, and allow users to override this with
force. As such, it does not make sense to have a changing default base
to deploy to, this should be hard coded as well.

Remove this fallback option from the BaseSelector, fallback straight to
the DefaultSupportedLTS