Domain errors
This masks all sql, sqlite and dqlite errors so they're not available
via errors.Is or errors.As. All errors should be trapped at the domain
layer and not leaked outside. Thus the API client and server are never
coupled to the underlying database errors.

Dropping the older domain errors for specific domain type errors gives
more focused errors that can be trapped at the API server layer.