Mapper watchers
Instead of using a predicate to define if a change event should be
filtered out, we can instead use a mapper. There was a flaw in with
the predicate that meant that if one event was tainted then all
events where tainted. This allows us to drop the tainted event whilst
returning all the other events.

Generalising this further we can then map a change event to another
event, or we can ask if a change event can be inserted, filtered or
dropped (think flatMap).