Add TestRevokeAccess.
Test the different permutations of revoking access to an object for the
access. Replaces some tests in the controller, model and offer facades
around revoking access there. These tests use to do a full round trip to
state. Now they are mocked and the actual functionality is in the Access
domain.