Swap state for AccessService.
The ControllerAccessService replaces ControllerAccess. Only two methods
are now required ReadUserAccessLevelForTarget and UpdatePermission.
UpdatePermission encapsulates the behavior of granting and revoking
controller permissions. Tests related to this behavior have been replaced
with mocked versions.

The controller tag will not change. No need to ask for it over and over,
keep the value when the API is created.