Rename model logger
Renaming Logger and ModelLogger allows us to reuse the Logger name
in core/logger package. This is a move to centeralise the logging
interfaces through the juju package. Unfortunately, we have multiple
Logger interfaces throughout the juju package, which is normally
ok if the function on the interface do not reference itself. If they
do reference itself (Child and ChildWithLabels), then we need to
insert a shim to convert it from one interface to anther because of
the invariance of types in go.

Using LogWriter instead of Logger indicates that it's expected to
write the log record using the Log method.