Create the basic Logger interface
For better or worse we just want a centeralised logging interface.
Without re-writing the whole of Juju we should just standardise on
what we have. We can then move to a better one at a later date.

I suspect we could even still use this interface and replace loggo
with slog at a later date.