Use a centralised logger interface
This standardises all of the logging in juju to one and only one
interface. This will ensure that when we implement status history
that we can guarantee that the logging is correctly nested and
can pass labels through.

Currently loggo is the backing for logger.Logger, but it should
be really easy to just change this to slog if so desired.