Drop series parsing from supported-bases code
Originally, ControllerBases and WorkloadBases were shims around
ControllerSeries and WorkloadBases

De-couple these functions from eachother

We still read distro-info, which will remain the case until after Madrid
whewre a final decision for direction can be made.

As such, we still read the codename, and do use this in a few places.
However, we treat the codenames as opaque strings, and never parse them
as OS systems. This is as close as we can get before Madrid