Swaps model config client facade to DQlite.
With this change we are now making the model config client facade use
DQlite backed services for reading, writing and removing model config
values.

All of the model config checking logic bar one item has been moved down
into the services layer. The remaining one requires permission checks so
we have left that as an API concern.

The storage and secret backend checkers need to be fully implemented as
well. This will be worked on with Wallyworld in a follow up PR.