Use channel in RefreshBase in charm revision updater
RefreshBases, at the moment, are agnostic to the format of their
channel. They can be channels, channels with risks, or series

In the charm revision updater for some reason we go with a series, which
is parsed from a charm origin in a previous step.

Use a channel instead which, as it turns out, was already available

See here, where the channel is finally read:
https://github.com/juju/juju/blob/3.6/charmhub/refresh.go#L425-L432

You see we first check if the channel is a series, and if it is not, in
the sanatiseChannel step, we parse it as a channel and extract the track