Fix returned errors on network domain state layer
This patch fixes two cases on which the wrong error was returned from
the state layer on the network domain, meaning that the service layer
wasn't handling those correctly.

The first one is when trying to add a subnet with the same provider-id
(this happens when reloading spaces twice on a row for example). The
second one is when trying to remove spaces that don't contain a provider
id, this should not prevent the delete.