Clean spaces/subnets legacy state usages and references
This patch removes state spaces and subnets, the collections, tests and
any dangling references left on the apiserver and the containerizer
worker.

Note: The model config checks regarding the default space has been
removed (since it was done on the agent bootstrap) and the new model
config backed with dqlite allows us to do this in the new modelconfig
domain. This will land in a following patch.