Remove check for machine addresses when updating endpoint bindings
This patch (temporarily) removes a check used when updating endpoint
bindings, that makes sure that the machine(s) actually do contain
addresses on the spaces that were passed or already existing as
bindings.

This is part of the cleanup of spaces and subnets in the legacy state
and therefore this check must be reimplemented when migrating endpoint
bindings to dqlite.