One source of truth for the UUID when creating a model.
Having a uuid as an argument and a uuid in the ModelCreationArgs gives
you 2 potential sources of truth when creating a model. This gets
confusing when coming into the code new. All of the callers, modulo
tests, have created a UUID for the model before creating it. Use the
UUID in the args.