Simplifies readonly model data creation.
We require the model metadata to available in the model database for
services scoped to a model database. This required a lot of work from
model creation components to duplicate data and handle the transfer from
one database to another.

By exposing both databases to the service we can make the service do the
heavy lifting and not push the complicated logic upwards. This allows
for additions like agent version to much more simplified.

This commit also exposed the agent version in the model metadata.