Simplifies fetching ModelConfig.
Because we now have access to the controller database and model uuid
inside of the model service we can construct the model config service
for the user without them having to build their own defaults service to
supply to the factory.

This will make it much cleaner and easier for users of the model config
service.