Drop support for series in RefreshBase
Up until now we supported constructing the RefreshBase type with a
series in the channel. However, no where in the code base do we make use
of this.

I checked every read and write to the Channel attribute and nowhere do
we fill in what could be a series.

Since we're moving away from the series construct, drop suport for this.