Remove unused region field in gce provider Connection
This is something we stumbled upon recently with @manadart when
figuring out https://bugs.launchpad.net/juju/+bug/2058714. The region
information in the environ is always picked up from the cloud spec, so
the region field in the Connection seems redundant.

There's more to be done in this piece of code to unravel some of the
unnecessary levels of abstractions. However, this small change is a
clear win for now.