Rename vsphere template dir
In 3.5 we would cache our images in directories named according to the
series. This meant converting a series to a base

Dop this conversion and instead name these directorys based on the
striaght base

Since 3.6 is a new minor version requiring model migration to access, we
don't need to worry about incompatibilities when upgrading in place

Fortunately, once a template is created, we only pass through the
template structure (i.e. the whole dir), so this change was fairly simple
and painless