Introduces an easier model config set for bootstrap.
This change now makes it so bootstrap model config takes raw values and
not a config object that requires specific values to be pre set. What
this allows us to do is have model config bootstrap correctly construct
a models config without the user having to understand what makes a valid
model config.

This is important for facade testing to work correctly as well.