Updates the agent version validator for config.
With the move to DQlite we want to stop using model config to store and
keep track of a models agent version. Instead we want this value to be a
first class citzen in both the database and Juju model.

However this change needs to be done in stages over the course of the
DQlite move so we still support reading agent version from model config.

We achieve this by adding it to the model config that is returned to a
caller at the last moment. However there still exists the chance that
this value could still be persisted into the database.

With this commit we are extending the agent version validator to remove
the attribute from model config updates so it can't make it down into
the state layer.