Makes the read only model service scoped.
The read only model service has been introduced so that a Juju
controller can inject metadata information about a model into the model
database and also provide an easy access method for model scoped facades
to request information about the model they are working on.

This service made an assumption in a few cases that the caller was
responsible for providing the id of the model when that information has
already been supplied to the scoped database. It introduces the
possability for error and data drift.

These changes remove the need to supply the model id and makes the
service far more scrict in what it will accept for model information.