Improve changestream pruner log output
The changestream pruner will log out if the changestream isn't keeping
up with all the terms in the changestream. Unfortunately, if there
aren't any writes going to the changestream it will still message out
that it isn't keeping up. This isn't actually true. There just isn't
enough changes being added to the changestream for the log message to
be useful. Refactoring it to only log out if the window of changes has
moved will then surface the real log messages.