Fix application API to add correct endpoint bindings to spaceIDs
Previously, we used to have a mix of spaceIDs and names, now with the
new network domain we need to make sure all bindings contain spaceIDs.
This patch adds the conversion from space names to IDs (which was already present for one of the
API endpoints) to every endpoint that creates the endpoint bindings.