Service factory at model worker level
Push the service factory through the model worker manager. As we
require the model config and ancillary services, we need the service
factory available to us in the model workers. As the model workers
are just hosted on the jujud-controller we can now just get at it
directly. There is no longer an API barrier to cross.

We already do this for the provider service factory, so we can
do the same here.