Move pruners over to service factory
The action and status history pruners are originally using the
model config api to get the information. We don't want to do that
any more. Instead we want to use the service factory directly. This
allows for a more efficent streamlined juju. We can in theory drop
API methods as we don't need to expose all of the internal calls
via the API server.

The work is simple, use the new service factory from the model worker
changes (see #17343) and just start to use the factory directly.