Drop support for series in bundles
Support for series in bundles is being dropped in 4.0 (major version
bump). Error out if bundles include series

This means we no longer need to parse series into base in the rest of
our bundle deploy code

We do not want to rely the juju/charm BundleData struct to parse series.
We want to keep this data structure clean, so will be removing from
BundleData very soon Series

As such, we need to parse the bundle ourselves. This requires using a
new miner version of "juju/charm" which allows us to access the raw
unparsed bundle

Also, swap our series for base in test bundles