Swap all loggo.GetLogger calls to internal version
Most if not all remaining calls to loggo package are now to setup
the writers. We can take care of that another time. It would be a
case of pushing that into the internal package and making the writers
have a basic standard.

For now, I've pushed this as much as I feel is required.