Fix logical race in tests
The original tests stated that the expected test case was done
when the pre-upgrade steps was complete. This wasn't exactly true.
Instead, the worker test would race to the kill the worker and we
could call the catacomb dying case.
The fix is to wait for the worker to die based on the steps worker
and return correctly. This will then return the correct error message.